requirements.txt
spyder
pyqt5-tools
matplotlib
scikit-image